Refactor financial calculations to use decimal types
Transitioned from `double` and `ulong` to `decimal` for improved precision in financial and resource calculations across multiple classes. Introduced a stockpile system in `City.cs` for managing resources like Food, Timber, Stone, and Iron, with relevant properties updated to `decimal`. Updated `Nations.cs`, `NationBonus.cs`, `NationRelation.cs`, and `PopulationUnit.cs` for consistency in numerical handling. Added new classes and methods, including `IBuildable` and `CityEconomy`, to enhance city economics management. Introduced the `Yield` struct for resource yields and implemented economic growth logic in `CityEconomyGrowthEvent.cs`. Added utility methods in `NumericalExtensions.cs` for better numerical operations.

+++ Core/Mechanics/Nations.cs
@@ -1,5 +1,6 @@
 ﻿using Civ2Like.Core.Nations;
 using Civ2Like.Core.World;
+using Civ2Like.Extensions;
 using Civ2Like.Hexagon;
 
 namespace Civ2Like.Core.Mechanics;
@@ -8,16 +9,16 @@ public static class Nations
 {
     private static readonly Random _rand = new();
 
-    private const double GROWTH_NOISE_MAX = 0.05;   // was 0.2 → ±0.5% in the final rate
-    private const double GROWTH_TERRAIN_SCALE = 4.0;
-    private const double GROWTH_RATE_SCALE = 0.10; // your original 0.1
+    private const decimal GROWTH_NOISE_MAX = 0.05m;   // was 0.2 → ±0.5% in the final rate
+    private const decimal GROWTH_TERRAIN_SCALE = 4.0m;
+    private const decimal GROWTH_RATE_SCALE = 0.10m; // your original 0.1
 
-    private const double MIGRATION_BASE = 0.02; // was 0.05 → ~2% baseline
-    private const double MIGRATION_NOISE_MAX = 0.05; // was 0.1
-    private const uint MIGRATION_MIN_MOVE = 1;   // don’t bother if < 10 people
+    private const decimal MIGRATION_BASE = 0.02m; // was 0.05 → ~2% baseline
+    private const decimal MIGRATION_NOISE_MAX = 0.05m; // was 0.1
+    private const decimal MIGRATION_MIN_MOVE = 1m;   // don’t bother if < 10 people
 
     // weights for attractiveness
-    private const double ATTR_TERRAIN_W = 1.0;
+    private const decimal ATTR_TERRAIN_W = 1.0m;
 
 
     public static void ApplyGrowth(Nation nation, IReadOnlyDictionary<Hex, Tile> map)
@@ -26,29 +27,29 @@ public static class Nations
         {
             PopulationUnit pu = pos.Value.Populations[nation];
 
-            double coefficient = TerrainFertility(pos.Value.Terrain) + _rand.NextDouble() * GROWTH_NOISE_MAX;
+            decimal coefficient = TerrainFertility(pos.Value.Terrain) + _rand.NextDecimal() * GROWTH_NOISE_MAX;
             coefficient += nation.Bonuses.Sum(i => i.GrowthMultiplier);
 
-            ulong growth = (ulong)(pu.Value * coefficient * GROWTH_RATE_SCALE);
+            long growth = (long)(pu.Value * coefficient * GROWTH_RATE_SCALE);
 
             pu.Value += growth;
         }
     }
 
-    private static double TerrainFertility(Terrain terrain) => (terrain switch
+    private static decimal TerrainFertility(Terrain terrain) => (terrain switch
     {
-        Terrain.Ocean => 0.00,
-        Terrain.Mountains => 0.00,
-        Terrain.Desert => 0.01,
-        Terrain.Snow => 0.005,
-        Terrain.Tundra => 0.01,
-        Terrain.Swamp => 0.012,
-        Terrain.Hills => 0.012,
-        Terrain.Forest => 0.015,
-        Terrain.Jungle => 0.015,
-        Terrain.Plains => 0.02,
-        Terrain.Grassland => 0.022,
-        Terrain.Coast => 0.0,
+        Terrain.Ocean => 0.00m,
+        Terrain.Mountains => 0.00m,
+        Terrain.Desert => 0.01m,
+        Terrain.Snow => 0.005m,
+        Terrain.Tundra => 0.01m,
+        Terrain.Swamp => 0.012m,
+        Terrain.Hills => 0.012m,
+        Terrain.Forest => 0.015m,
+        Terrain.Jungle => 0.015m,
+        Terrain.Plains => 0.02m,
+        Terrain.Grassland => 0.022m,
+        Terrain.Coast => 0.0m,
         _ => throw new NotImplementedException(),
     }) * GROWTH_TERRAIN_SCALE;
 
@@ -59,15 +60,15 @@ public static class Nations
         foreach (var pos in originalState.Where(i => i.Value.Populations.ContainsKey(nation)))
         {
             PopulationUnit pu = pos.Value.Populations[nation];
-            double coefficient = 1.0 + TerrainMigrationability(pos.Value.Terrain) + _rand.NextDouble() * MIGRATION_NOISE_MAX;
+            decimal coefficient = 1.0m + TerrainMigrationability(pos.Value.Terrain) + _rand.NextDecimal() * MIGRATION_NOISE_MAX;
 
             coefficient += nation.Bonuses.Sum(i => i.MigrationMultiplier);
 
-            ulong allMigratingPeople = (ulong)(pu.Value * coefficient * MIGRATION_BASE);
+            decimal allMigratingPeople = pu.Value * coefficient * MIGRATION_BASE;
 
             var neighbours = map.Neighbors(pos.Key).
                 Select(i => map[i]).
-                Where(i => TerrainFertility(i.Terrain) > 0.0).
+                Where(i => TerrainFertility(i.Terrain) > 0.0m).
                 OrderBy(_ => _rand.Next());
 
             foreach (var tile in neighbours)
@@ -77,8 +78,8 @@ public static class Nations
                     break;
                 }
 
-                double desirability = TerrainFertility(tile.Terrain); 
-                uint migratingPeople = (uint)(allMigratingPeople * desirability * 10);
+                decimal desirability = TerrainFertility(tile.Terrain);
+                decimal migratingPeople = allMigratingPeople * desirability * 10;
 
                 if (migratingPeople > 0)
                 {
@@ -104,20 +105,20 @@ public static class Nations
         }
     }
 
-    private static double TerrainMigrationability(Terrain terrain) => (terrain switch
+    private static decimal TerrainMigrationability(Terrain terrain) => (terrain switch
     {
-        Terrain.Ocean => 0.00,
-        Terrain.Mountains => 0.00,
-        Terrain.Desert => 0.01,
-        Terrain.Snow => 0.005,
-        Terrain.Tundra => 0.01,
-        Terrain.Swamp => 0.012,
-        Terrain.Hills => 0.012,
-        Terrain.Forest => 0.015,
-        Terrain.Jungle => 0.015,
-        Terrain.Plains => 0.02,
-        Terrain.Grassland => 0.022,
-        Terrain.Coast => 0.0,
+        Terrain.Ocean => 0.00m,
+        Terrain.Mountains => 0.00m,
+        Terrain.Desert => 0.01m,
+        Terrain.Snow => 0.005m,
+        Terrain.Tundra => 0.01m,
+        Terrain.Swamp => 0.012m,
+        Terrain.Hills => 0.012m,
+        Terrain.Forest => 0.015m,
+        Terrain.Jungle => 0.015m,
+        Terrain.Plains => 0.02m,
+        Terrain.Grassland => 0.022m,
+        Terrain.Coast => 0.0m,
         _ => throw new NotImplementedException(),
     }) * ATTR_TERRAIN_W;
 }
